pyproject.toml
[project]
name = "inkpotro"
version = "0.1.0"
description = "Write Freely, Publish Instantly."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "markdown>=3.8.2",
    "pygithub>=2.6.1",
    "pyinstaller>=6.14.0",
    "pyqt6>=6.9.1",
]
